+++ SalonDeMasajes/src/sdm/repositorios/Dao.java
@@ -3,11 +3,11 @@ package sdm.repositorios;
 public interface Dao<T> {
 	Iterable<T> obtenerTodos();
 
-	T obtenerPorId(Long id);
+	T obtenerPorId(Integer id);
 
 	void agregar(T objeto);
 
 	void modificar(T objeto);
 
-	void borrar(Long id);
+	void borrar(Integer id);
 }
